+++ src/org/omegat/util/gui/FontFallbackListener.java
@@ -4,6 +4,7 @@
           glossaries, and translation leveraging into updated projects.
 
  Copyright (C) 2015 Aaron Madlon-Kay
+               2021 Hiroshi Miura
                Home page: http://www.omegat.org/
                Support center: https://omegat.org/support
 
@@ -41,6 +42,8 @@ import javax.swing.text.SimpleAttributeSet;
 import javax.swing.text.StyleConstants;
 import javax.swing.text.StyledDocument;
 
+import org.omegat.util.Preferences;
+
 
 public class FontFallbackListener implements DocumentListener {
 
@@ -63,6 +66,9 @@ public class FontFallbackListener implements DocumentListener {
     }
 
     private void doStyling(Document document, final int offset, final int length) {
+        if (!Preferences.isPreference(Preferences.FONT_FALLBACK)) {
+            return;
+        }
         if (!(document instanceof StyledDocument)) {
             return;
         }
